Fixed bug in clearWinConFig util methods
clearWinConFig util methods would assign an int into a board piece slot instead of BoardPiece.m_piecePostion. Causing a new board piece to be constructed making the methods appear to be working correctly.  They now call reset() on the board piece to properly reset it without losing the piece position on the board, or creating an entirly
new board slot with the same position.